Removing unneeded files and directories, also bringing project to the standard structure.     Resource loading was modified to load image from classpath. Please review and ask questions, you should have some :-)
