--- .IdeaIC2016.2/config/options/ignore.xml
@@ -1,34 +1,0 @@
-<application>
-  <component name="IgnoreSettings" missingGitignore="true" ignoredFileStatus="true" outerIgnoreRules="true" outerIgnoreWrapperHeight="100" version="1.5" starredTemplates="" settingsFixed="true">
-    <languages>
-      <language id="Bazaar" NEW_FILE="true" ENABLE="true" />
-      <language id="CloudFoundry" NEW_FILE="true" ENABLE="false" />
-      <language id="Chef" NEW_FILE="true" ENABLE="false" />
-      <language id="Cvs" NEW_FILE="true" ENABLE="true" />
-      <language id="Darcs" NEW_FILE="true" ENABLE="true" />
-      <language id="Docker" NEW_FILE="true" ENABLE="false" />
-      <language id="ESLint" NEW_FILE="true" ENABLE="false" />
-      <language id="Floobits" NEW_FILE="true" ENABLE="false" />
-      <language id="Fossil" NEW_FILE="true" ENABLE="true" />
-      <language id="Git" NEW_FILE="true" ENABLE="true" />
-      <language id="Git exclude" NEW_FILE="true" ENABLE="true" />
-      <language id="Jetpack" NEW_FILE="true" ENABLE="false" />
-      <language id="JSHint" NEW_FILE="true" ENABLE="false" />
-      <language id="Mercurial" NEW_FILE="true" ENABLE="true" />
-      <language id="Monotone" NEW_FILE="true" ENABLE="true" />
-      <language id="Nodemon" NEW_FILE="true" ENABLE="false" />
-      <language id="Npm" NEW_FILE="true" ENABLE="false" />
-      <language id="Perforce" NEW_FILE="true" ENABLE="true" />
-      <language id="Team Foundation" NEW_FILE="true" ENABLE="true" />
-    </languages>
-    <userTemplates>
-      <template name="Example user template">### Example user template
-
-# IntelliJ project files
-.idea
-*.iml
-out
-gen</template>
-    </userTemplates>
-  </component>
-</application>
