--- .IdeaIC2016.2/config/options/other.xml
@@ -1,5 +1,0 @@
-<application>
-  <component name="RefactoringSettings">
-    <option name="MOVE_PREVIEW_USAGES" value="false" />
-  </component>
-</application>
