--- .IdeaIC2016.2/config/options/security.xml
@@ -1,16 +1,0 @@
-<application>
-  <component name="PasswordDatabase">
-    <option name="PASSWORDS">
-      <array>
-        <array>
-          <option value="133b77f6f01e93b6ce1d608a6f9f1fb6acb11297498f91719ea99ea448fd3caf" />
-          <option value="18df3364fc43d358f07f5ea711e4d44d5f3fbfddec7f1037b08fff55749373a4" />
-        </array>
-        <array>
-          <option value="b7b7085d1cf4089a4c73cf90ad6b093e" />
-          <option value="75297cbc5bc19729ccf3480286c9097d6f245b1a6693a7230f8e6e365c81ea0d" />
-        </array>
-      </array>
-    </option>
-  </component>
-</application>
