--- .IdeaIC2016.2/config/options/vcs.xml
@@ -1,12 +1,0 @@
-<application>
-  <component name="GitRememberedInputs">
-    <option name="visitedUrls">
-      <list>
-        <UrlAndUserName>
-          <option name="url" value="http://github.com" />
-          <option name="userName" value="bkhor" />
-        </UrlAndUserName>
-      </list>
-    </option>
-  </component>
-</application>
