+++ .gitignore
@@ -1,0 +1,6 @@
+target
+*.iml
+*.iws
+.idea
+*~
+.*~
