--- ME/java/Tanks/.idea/.gitignore
@@ -1,22 +1,0 @@
-extra/
-.idea/
-/.bundle
-/db/*.sqlite3
-/doc/
-*.rbc
-*.sassc
-.sass-cache
-capybara-*.html
-.rspec
-/vendor/bundle
-/log/*
-/tmp/*
-/public/system/*
-/coverage/
-/spec/tmp/*
-**.orig
-rerun.txt
-pickle-email-*.html
-/config/environments/production.rb
-/config/*.yml
-/*.log
