--- ME/java/Tanks/.idea/modules.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="ProjectModuleManager">
-    <modules>
-      <module fileurl="file://$PROJECT_DIR$/Res/Res.iml" filepath="$PROJECT_DIR$/Res/Res.iml" />
-      <module fileurl="file://$PROJECT_DIR$/Tanks.iml" filepath="$PROJECT_DIR$/Tanks.iml" />
-    </modules>
-  </component>
-</project>
