--- ME/java/Tanks/.idea/vcs.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="VcsDirectoryMappings">
-    <mapping directory="$USER_HOME$" vcs="Git" />
-  </component>
-</project>
