--- ME/java/Tanks/src/game/EntityType.java
@@ -1,8 +1,0 @@
-package game;
-
-
-public enum EntityType
-{
-    Player;
-
-}
