--- ME/java/Tanks/src/graphics/TextureAtlas.java
@@ -1,19 +1,0 @@
-package graphics;
-
-
-import utils.ResLoader;
-
-import java.awt.image.BufferedImage;
-
-public class TextureAtlas
-{
-    BufferedImage image;
-
-    public TextureAtlas(String imageName) {
-        image = ResLoader.loadImage(imageName);
-    }
-
-    public BufferedImage cut (int x , int y , int w , int h){
-        return image.getSubimage(x, y, w, h);
-    }
-}
