--- ME/java/Tanks/src/main/Main.java
@@ -1,16 +1,0 @@
-package main;
-
-import game.Game;
-
-
-
-public class Main
-{
-    public static void main(String[]args)
-    {
-        Game tanks = new Game();
-        tanks.start();
-
-
-    }
-}
