--- ME/java/Tanks/src/utils/Time.java
@@ -1,12 +1,0 @@
-package utils;
-
-
-public class Time
-{
-    public static final long  SECOND = 10000000000l;
-
-    public static long get()
-    {
-        return System.nanoTime();
-    }
-}
