+++ README.md
@@ -1,3 +1,9 @@
 #Tanks project 
 #Was made by Bohdan Khoruzhenko
-#Original game was Tanks Battlecity 
+#Original game was Tanks Battlecity
+
+
+To run enter:
+
+mvn clean compile exec:java
+
