+++ src/main/java/bkhor/games/tanks/game/Entity.java
@@ -1,0 +1,26 @@
+package bkhor.games.tanks.game;
+
+
+import java.awt.*;
+
+import bkhor.games.tanks.io.Input;
+
+public abstract class Entity
+{
+    public final EntityType type;
+
+    protected float x;
+    protected float y;
+
+    protected Entity(EntityType type, float x , float y)
+    {
+        this.type = type;
+        this.x = x;
+        this.y = y;
+
+    }
+
+    public abstract void update(Input input);
+
+    public abstract void render(Graphics2D g);
+}
