+++ src/main/java/bkhor/games/tanks/game/EntityType.java
@@ -1,0 +1,8 @@
+package bkhor.games.tanks.game;
+
+
+public enum EntityType
+{
+    Player;
+
+}
