+++ src/main/java/bkhor/games/tanks/io/Input.java
@@ -1,0 +1,49 @@
+package bkhor.games.tanks.io;
+
+
+import javax.swing.*;
+import java.awt.event.ActionEvent;
+import java.lang.reflect.Array;
+import java.util.Arrays;
+
+public class Input extends JComponent
+{
+    private boolean[] map;
+
+    public Input() {
+        map = new boolean[256];
+
+        for(int i=0; i<map.length; i++)
+        {
+
+            final int KEY_CODE = i;
+
+            getInputMap(JComponent.WHEN_IN_FOCUSED_WINDOW).put(KeyStroke.getKeyStroke(i, 0, false), i * 2);
+            getActionMap().put(i * 2, new AbstractAction() {
+                @Override
+                public void actionPerformed(ActionEvent actionEvent) {
+                    map[KEY_CODE] = true;
+                }
+            });
+
+            getInputMap(JComponent.WHEN_IN_FOCUSED_WINDOW).put(KeyStroke.getKeyStroke(i, 0, true), i * 2 + 1);
+            getActionMap().put(i * 2 + 1, new AbstractAction() {
+                @Override
+                public void actionPerformed(ActionEvent actionEvent) {
+                    map[KEY_CODE] = false;
+                }
+            });
+        }
+    }
+
+    public boolean[] getMap() {
+        return Arrays.copyOf(map , map.length);
+
+    }
+
+    public boolean getKey(int keyCode){
+        return map[keyCode];
+
+    }
+
+}
