+++ src/main/java/bkhor/games/tanks/main/Main.java
@@ -1,0 +1,16 @@
+package bkhor.games.tanks.main;
+
+import bkhor.games.tanks.game.Game;
+
+
+
+public class Main
+{
+    public static void main(String[]args)
+    {
+        Game tanks = new Game();
+        tanks.start();
+
+
+    }
+}
