+++ src/main/resources/bkhor/games/tanks/utils/texture_atlas.png
